seperated controllers, and tests into individual controller directory, and removed controller index.js
